requirements.txt
# ELK Auto Resolver 필수 패키지
elasticsearch==8.5.0
kubernetes==24.2.0
psycopg2-binary==2.9.5
PyYAML==6.0
requests==2.28.2
python-dotenv==0.19.2
psutil==5.9.4
urllib3==1.26.18

# 추가 유틸리티 패키지
setuptools>=65.0.0
wheel>=0.37.0